[chore] Use int64 for queue_size in the queuebatch.Config, simplifies conversions (#12745)

+++ exporter/exporterhelper/internal/queue_sender.go
@@ -58,7 +58,7 @@ type QueueConfig struct {
 	Sizer request.SizerType `mapstructure:"sizer"`
 
 	// QueueSize represents the maximum data size allowed for concurrent storage and processing.
-	QueueSize int `mapstructure:"queue_size"`
+	QueueSize int64 `mapstructure:"queue_size"`
 
 	// NumConsumers is the number of consumers from the queue.
 	NumConsumers int `mapstructure:"num_consumers"`
